Made acquire/release system for In/Out streams consistant, and fixed execution space issue in c++

+++ generic-cpp-receiver/src/main/cpp/RCode/commands/RCodeCapabilitiesCommand.cpp
@@ -11,17 +11,18 @@ void RCodeCapabilitiesCommand::execute(RCodeCommandSlot *slot,
         RCodeCommandSequence *sequence, RCodeOutStream *out) {
     if (RCodeParameters::bigBigFieldLength > 0) {
         RCodeOutStream::writeFieldType(out, 'B',
-                RCodeParameters::bigBigFieldLength);
+                (uint32_t) RCodeParameters::bigBigFieldLength);
     } else {
         RCodeOutStream::writeFieldType(out, 'B',
-                RCodeParameters::bigFieldLength);
+                (uint32_t) RCodeParameters::bigFieldLength);
     }
     uint8_t capabilities = 0x02;
     if (RCodeParameters::slotNum > 1) {
         capabilities |= 0x04;
     }
-    RCodeOutStream::writeFieldType(out, 'N', RCodeParameters::fieldNum);
-    RCodeOutStream::writeFieldType(out, 'M', 1);
+    RCodeOutStream::writeFieldType(out, 'N',
+            (uint32_t) RCodeParameters::fieldNum);
+    RCodeOutStream::writeFieldType(out, 'M', (uint8_t) 1);
     out->writeField('U', sizeof(fieldUnit));
     RCodeSupportedCommandArray supported =
             rcode->getCommandFinder()->getSupportedCommands();
